Restructure AuxInfo: don't rely on `parent id` serialized in AuxTree
Node `parent id` is a redundant information because each inner node already
contains ids of left and right nodes. Parent id can be derived from
the existing data.

This is a preparation for re-using of the `parent id` field to store
different information without increasing the size of the AuxTrees.